command SEX and AVERAGE implementation, fixed patient data
Both commands work and are fully working.
Patient data now works and shows correctly.

+++ patient.h
@@ -2,13 +2,13 @@
 #include "date.h"
 
 typedef struct patient{
-    int id;
+    long int id;
     char sex[6]; // {“male”, “female”}
     int birthYear;
     char country[40]; //birth country
     char region[40];
     char infectionReason[100];
-    int infectedBy; // id of the infected patient
+    long int infectedBy; // id of the infected patient
     Date confirmedDate;
     Date releasedDate;
     Date deceasedDate;
@@ -17,7 +17,7 @@ typedef struct patient{
 
 typedef struct patient *PtPatient;
 
-Patient createPatient(int id, char *sex, int birthYear, char* country, char* region, char* infectionReason, int infectedBy, Date confirmedDate, Date releasedDate, Date deceasedDate, char* status);
+Patient createPatient(long int id, char *sex, int birthYear, char* country, char* region, char* infectionReason, long int infectedBy, Date confirmedDate, Date releasedDate, Date deceasedDate, char* status);
 void printPatient(Patient p);
 int getAge(Patient p);
 int getNumberOfInfectedDays(Patient p);
